requirements.txt
django>=5.1.1
djangorestframework>=3.15.2
environs>=11.0.0
gunicorn>=23.0.0